Diffuse error in H direction on scr5
Screens converted in screen 5 don't seem to produce checkerboards, only alternate lines. I think that's because the error is not being spread in horizontal direction. Maybe this x+1 should have been x+2 instead? This error doesn't happen in scr2, but may be present in other screens. I have no php installed so this is untested.